Extract file policy, terminal sessions/execution, OCI effective config, and local runtime owners (#82591)
